test: move tests for deleteMessagesFromConversation utility function to relevant file
